+++ HISTORY.rst
@@ -7,6 +7,7 @@ History
 
 * Better handling for a large variety of error conditions
 * Proper fallback mechanism implemented
+* Added ``--watch`` option.
 * Various bug fixes:
   * Avoid output ``.elm`` files with no exports
   * Bugs with HTML attributes and non-string message args